% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_spatial_sse}
\alias{compute_spatial_sse}
\title{Compute the Sum of Squared Errors (SSE) for the 2D Spatial Piecewise Model}
\usage{
compute_spatial_sse(observed, baseline, latent, beta, c)
}
\description{
Compute the Sum of Squared Errors (SSE) for the 2D Spatial Piecewise Model
}
