% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_basis.R
\name{construct_rbf}
\alias{construct_rbf}
\title{Generate Radial Basis Functions (Thin-plate splines)}
\usage{
construct_rbf(x, y, knots_x, knots_y)
}
\arguments{
\item{x}{Numeric vector of x coordinates}

\item{y}{Numeric vector of y coordinates}

\item{knots_x}{Numeric vector of knot x coordinates}

\item{knots_y}{Numeric vector of knot y coordinates}
}
\value{
A matrix of radial basis evaluations
}
\description{
Generates the underlying smooth spatial surface Z(x,y) from a set of knots.
In a Bayesian framework, the weights of these basis functions are estimated
to warp the boundary contours around the protein puncta.
}
