% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_landscape.R
\name{distance_to_segment}
\alias{distance_to_segment}
\title{Calculate shortest distance to a line segment}
\usage{
distance_to_segment(px, py, ax, ay, bx, by)
}
\description{
Calculate shortest distance to a line segment
}
\keyword{internal}
