% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialbp_formula.R
\name{hinge}
\alias{hinge}
\title{Define a spatial hinge term for the formula interface}
\usage{
hinge(latent, mod_beta = NULL, mod_c = NULL, smooth = FALSE)
}
\arguments{
\item{latent}{The latent spatial gradient variable}

\item{mod_beta}{One-sided formula specifying covariates that modify the slope}

\item{mod_c}{One-sided formula specifying covariates that modify the breakpoint}

\item{smooth}{Logical. If TRUE, estimates a smooth transition (Softplus rho)}
}
\description{
Define a spatial hinge term for the formula interface
}
