% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_basis.R
\name{level_set_mask}
\alias{level_set_mask}
\title{Spatial Mask (Discontinuous Level Set)}
\usage{
level_set_mask(z, c)
}
\arguments{
\item{z}{Numeric vector representing the latent spatial propensity field.}

\item{c}{Numeric scalar, the breakpoint/threshold contour.}
}
\value{
An integer vector (0 or 1) spatial mask.
}
\description{
A stricter boundary where the surface is allowed to abruptly jump
(discontinuous) rather than smoothly hinge.
}
