% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialbp_model.R
\name{plot.spatialbp}
\alias{plot.spatialbp}
\title{Plot a 2D Spatial Piecewise Regression}
\usage{
\method{plot}{spatialbp}(x, ...)
}
\arguments{
\item{x}{An object of class spatialbp}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Intercepts the generic plot() call to prevent smoothbp from trying to draw
a 1D line. Instead, this renders a 2D topographical map of the boundaries.
}
