% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{run_joint_mcmc}
\alias{run_joint_mcmc}
\title{Fit a Joint 2D Spatial Piecewise Regression Model}
\usage{
run_joint_mcmc(
  y1,
  y2,
  baseline1,
  baseline2,
  latent,
  mode,
  iterations,
  init_beta1,
  init_beta2,
  init_c,
  sd_beta1,
  sd_beta2,
  sd_c,
  lambda1,
  lambda2
)
}
\description{
Supports both "affinity" (shared boundary, same direction) and
"repulsion" (shared boundary, opposite directions) hypotheses.
}
