% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{run_spatial_mcmc}
\alias{run_spatial_mcmc}
\title{Run a Random Walk Metropolis-Hastings MCMC to fit the spatial breakpoint}
\usage{
run_spatial_mcmc(
  observed,
  baseline,
  latent,
  iterations,
  init_beta,
  init_c,
  sd_beta,
  sd_c,
  lambda
)
}
\arguments{
\item{observed}{The observed pixel intensities (multiplexed histology data)}

\item{baseline}{The baseline spatial surface Z(x,y)}

\item{latent}{The latent propensity field}

\item{iterations}{Number of MCMC iterations}

\item{init_beta}{Starting value for beta}

\item{init_c}{Starting value for c}

\item{sd_beta}{Proposal standard deviation for beta}

\item{sd_c}{Proposal standard deviation for c}

\item{lambda}{L1 (Laplace) regularization parameter for beta}
}
\description{
Run a Random Walk Metropolis-Hastings MCMC to fit the spatial breakpoint
}
