% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialbp_formula.R
\name{spatialbp}
\alias{spatialbp}
\title{Fit a Spatial Piecewise Model via Formula}
\usage{
spatialbp(
  formula,
  data,
  mode = "independent",
  iterations = 5000,
  warmup = floor(iterations/2),
  sd_beta = 0.1,
  sd_c = NULL
)
}
\arguments{
\item{formula}{A formula object (e.g., y ~ x + offset(z) + hinge(latent_z))}

\item{data}{A data frame}

\item{mode}{Character, "independent", "affinity", "repulsion"}

\item{iterations}{Integer, number of MCMC iterations}
}
\description{
Fit a Spatial Piecewise Model via Formula
}
