% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity_test.R
\name{test_spatialbp}
\alias{test_spatialbp}
\title{Multivariate Spatial Affinity Test}
\usage{
test_spatialbp(red, green, blue, latent_field, iterations = 10000)
}
\arguments{
\item{red}{Numeric vector of Red protein intensities}

\item{green}{Numeric vector of Green protein intensities}

\item{blue}{Numeric vector of Blue protein intensities}

\item{latent_field}{Precomputed spatial radial basis values}

\item{iterations}{Number of MCMC iterations}
}
\value{
A list containing Bayes Factors and posterior probabilities.
}
\description{
Tests the affinity and repulsion hypotheses between three protein stains
(Red, Green, Blue) using the spatialbp Rust MCMC engine.
}
