% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialbp_model.R
\name{plot.spatialbp}
\alias{plot.spatialbp}
\title{Plot a 2D Spatial Piecewise Regression}
\usage{
\method{plot}{spatialbp}(x, z_slice = NULL, t_slice = NULL, facet_by = NULL, ...)
}
\arguments{
\item{z_slice}{Numeric (optional), specific Z-slice to display for 3D/4D data.}

\item{t_slice}{Numeric (optional), specific time-point to display for 4D data.}

\item{facet_by}{Character (optional), column name to facet by.}

\item{...}{Unused.}
}
\description{
Intercepts the generic plot() call to prevent smoothbp from trying to draw
a 1D line. Instead, this renders a 2D topographical map of the boundaries.
}
